Revert "feat(integrations): add Intelligence threads to mcp-apps (#5211)"
This reverts commit bc8cec60920258b829324c2b52ca9329e1aeec81, reversing
changes made to 2103d0c875c7c8fa695f7f88dd1e5f7e133b6f65.
