--- examples/integrations/mcp-apps/.env.example
@@ -1,7 +1,0 @@
-OPENAI_API_KEY=your-openai-api-key-here
-
-# Optional: enable CopilotKit Intelligence Threads locally
-COPILOTKIT_LICENSE_TOKEN=
-INTELLIGENCE_API_KEY=
-INTELLIGENCE_API_URL=http://localhost:4201
-INTELLIGENCE_GATEWAY_WS_URL=ws://localhost:4401
