--- examples/integrations/mcp-apps/app/components/threads-drawer/index.tsx
@@ -1,4 +1,0 @@
-"use client";
-
-export { default as ThreadsDrawer } from "./threads-drawer";
-export type { ThreadsDrawerProps } from "./threads-drawer";
