+++ examples/integrations/mcp-apps/app/globals.css
@@ -3,31 +3,6 @@
 :root {
   --background: #ffffff;
   --foreground: #171717;
-
-  --card: oklch(1 0 0);
-  --card-foreground: oklch(0.145 0 0);
-  --primary: oklch(0.205 0 0);
-  --primary-foreground: oklch(0.985 0 0);
-  --secondary: oklch(0.97 0 0);
-  --secondary-foreground: oklch(0.205 0 0);
-  --muted: oklch(0.97 0 0);
-  --muted-foreground: oklch(0.556 0 0);
-  --accent: oklch(0.97 0 0);
-  --accent-foreground: oklch(0.205 0 0);
-  --destructive: oklch(0.577 0.245 27.325);
-  --destructive-foreground: oklch(0.985 0 0);
-  --border: oklch(0.922 0 0);
-  --input: oklch(0.922 0 0);
-  --ring: oklch(0.708 0 0);
-
-  --radius: 0.625rem;
-  --radius-sm: calc(var(--radius) - 4px);
-  --radius-md: calc(var(--radius) - 2px);
-  --radius-lg: var(--radius);
-  --radius-xl: calc(var(--radius) + 4px);
-
-  --font-body: Arial, Helvetica, sans-serif;
-  --font-code: "SFMono-Regular", Menlo, monospace;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -41,12 +16,6 @@ body {
   background: var(--background);
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
-}
-
-.threadsLayout,
-body > [role="presentation"] {
-  --foreground: oklch(0.145 0 0);
-  --background: oklch(1 0 0);
 }
 
 body,
