+++ examples/integrations/mcp-apps/next.config.ts
@@ -2,11 +2,6 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   serverExternalPackages: ["@copilotkit/runtime"],
-  env: {
-    NEXT_PUBLIC_COPILOTKIT_THREADS_ENABLED: process.env.COPILOTKIT_LICENSE_TOKEN
-      ? "true"
-      : "false",
-  },
 };
 
 export default nextConfig;
