+++ examples/integrations/mcp-apps/package.json
@@ -12,13 +12,10 @@
     "postinstall": "npm run install:mcp"
   },
   "dependencies": {
-    "@ag-ui/client": "0.0.53",
     "@ag-ui/mcp-apps-middleware": "^0.0.3",
-    "@copilotkit/react-core": "1.59.1",
-    "@copilotkit/react-ui": "1.59.1",
-    "@copilotkit/runtime": "1.59.1",
-    "hono": "^4.12.10",
-    "lucide-react": "^0.561.0",
+    "@copilotkit/react-core": "1.52.1",
+    "@copilotkit/react-ui": "1.52.1",
+    "@copilotkit/runtime": "1.52.1",
     "next": "16.1.1",
     "react": "^19.2.3",
     "react-dom": "^19.2.3"
