+++ examples/integrations/mcp-apps/tsconfig.json
@@ -29,5 +29,5 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts"
   ],
-  "exclude": ["node_modules", "threejs-server"]
+  "exclude": ["node_modules"]
 }
